requirements.txt
langchain
requests
python-docx
python-dotenv